feat: allow configuring the host port
Allows configuring the host port.

Normally this port should be left as the default port,
which is port 22. However, it may be that the ssh server is being run
without root, has a port above 1024.

Currently, this is a CLI option only using `--host-port PORT`.
There is no way to configure this by modifying `~/.ssh/config`.